feature(firebase): OBM-2818 Create Push Notification API. Put loadBalanced RestTemplate to the LEP context.
