+++ src/main/java/com/icthh/xm/tmf/ms/communication/config/RestTemplateConfiguration.java
@@ -1,5 +1,7 @@
 package com.icthh.xm.tmf.ms.communication.config;
 
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.cloud.client.loadbalancer.RestTemplateCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
@@ -11,4 +13,19 @@ public class RestTemplateConfiguration {
     public RestTemplate restTemplate() {
         return new RestTemplate();
     }
+
+    @Bean
+    @Qualifier("loadBalancedRestTemplate")
+    public RestTemplate loadBalancedRestTemplate(RestTemplateCustomizer customizer) {
+        RestTemplate restTemplate = new RestTemplate();
+        customizer.customize(restTemplate);
+        return restTemplate;
+    }
+
+    @Bean
+    @Qualifier("vanillaRestTemplate")
+    public RestTemplate vanillaRestTemplate() {
+        return new RestTemplate();
+    }
+
 }
